Expanded documentation. Refactored IKDW interface and implementation. Moved SetConfig intent creation into it's own class. Changed ScannerController abstract class so it uses property injection instead of constructor injection in order for custom implementations to actually be possible.
