+++ kotlin-data-wedge-lib/src/main/java/dk/gls/kdw/configuration/AppConfiguration.kt
@@ -4,6 +4,11 @@ import android.os.Bundle
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
+/**
+ * The app configuration configures which packageName the given configuration should apply too
+ * @param [packageName] The package name of the app the configuration should be enabled for
+ * @param [activityList] The activities in the app the configuration should be enabled in. Default = listOf(*), which enables it in the entire application.
+ */
 data class AppConfiguration(
     val packageName: String,
     val activityList: List<String> = listOf("*")
