+++ kotlin-data-wedge-lib/src/main/java/dk/gls/kdw/configuration/model/Plugin.kt
@@ -1,5 +1,8 @@
 package dk.gls.kdw.configuration.model
 
+/**
+ * The available plugins that are available for configuration using the library as of now
+ */
 enum class Plugin {
     BARCODE,
     INTENT
