+++ kotlin-data-wedge-lib/src/main/java/dk/gls/kdw/configuration/plugin/PluginConfiguration.kt
@@ -3,7 +3,13 @@ package dk.gls.kdw.configuration.plugin
 import android.os.Bundle
 import dk.gls.kdw.configuration.model.Plugin
 
-open class PluginConfiguration(
+/**
+ * The plugin to be configured
+ * @param [pluginName] the name of the plugin to configure
+ * @param [paramList] a bundle of the settings that should be applied
+ * @param [configModeReset] applies to an existing Profile with CONFIG_MODE CREATE_IF_NOT_EXIST or UPDATE. True (Default) – Clear any existing configuration and create a new configuration with the specified parameter values. False – Merge existing configuration with changes from the new configuration - update the existing values and add values not already in the configuration
+ */
+abstract class PluginConfiguration(
     val pluginName: Plugin,
     val paramList: Bundle,
     open val configModeReset: Boolean = true
