+++ kotlin-data-wedge-lib/src/main/java/dk/gls/kdw/model/intent/SubscriptionConfiguration.kt
@@ -2,6 +2,11 @@ package dk.gls.kdw.model.intent
 
 import android.os.Bundle
 
+/**
+ * The subscription configuration defines how data is delivered to the subscribing app
+ * [enabled] Whether the configuration is enabled
+ * [intentDelivery] Defines how the data should be delivered to the app. Default [IntentDelivery.BROADCAST]. BE AWARE If this is changed, the [ScannerController] on the KDW instance does not work as intended.
+ */
 data class SubscriptionConfiguration(
     val enabled: Boolean = true,
     val intentDelivery: IntentDelivery = IntentDelivery.BROADCAST,
