+++ kotlin-data-wedge-lib/src/main/java/dk/gls/kdw/model/label/LabelType.kt
@@ -1,5 +1,9 @@
 package dk.gls.kdw.model.label
 
+/**
+ * A convenience enum class containing known LabelType's
+ * Use the [String.toLabelType] function in order to interpret the incoming [ScannerResult] symbology as a [LabelType]
+ */
 enum class LabelType {
     DATAMATRIX,
     I2OF5,
@@ -8,8 +12,11 @@ enum class LabelType {
     UNKNOWN
 }
 
-fun LabelType.fromString(value: String) : LabelType {
-    return when(value) {
+/**
+ * Converts the received [ScannerResult] symbology string into a [LabelType]
+ */
+fun String?.toLabelType() : LabelType {
+    return when(this) {
         "LABEL-TYPE-DATAMATRIX" -> LabelType.DATAMATRIX
         "LABEL-TYPE-I2OF5" -> LabelType.I2OF5
         "LABEL-TYPE-CODE39" -> LabelType.CODE39
